requirements.txt
asgiref==3.6.0
cloudinary==1.32.0
dj-database-url==1.2.0
dj-rest-auth==3.0.0
Django==3.2.18
django-allauth==0.52.0
django-cloudinary-storage==0.3.0
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
oauthlib==3.2.2
Pillow==9.4.0
psycopg2==2.9.5
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.7.1
requests-oauthlib==1.3.1
sqlparse==0.4.3
types-cryptography==3.3.23.2
